Streaming: refactor StreamingHttpClient to a separate file
